Don't create zero length fallbacks unless necessary.
It’s not clear how fallbacks should be prioritised. It seems that
shorter fallbacks should be preferred over longer fallbacks to minimise
the amount of error; the exception seems to be when the length of the
fallback is zero, as that what essentially mean to insert a new
terminal into the input stream.